// src/api/models/MapData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MapGeneralWL } from './MapGeneralWL';
import {
    MapGeneralWLFromJSON,
    MapGeneralWLFromJSONTyped,
    MapGeneralWLToJSON,
    MapGeneralWLToJSONTyped,
} from './MapGeneralWL';
import type { MapPlayerWL } from './MapPlayerWL';
import {
    MapPlayerWLFromJSON,
    MapPlayerWLFromJSONTyped,
    MapPlayerWLToJSON,
    MapPlayerWLToJSONTyped,
} from './MapPlayerWL';

/**
 * 
 * @export
 * @interface MapData
 */
export interface MapData {
    /**
     * 
     * @type {string}
     * @memberof MapData
     */
    mapName: string;
    /**
     * 
     * @type {number}
     * @memberof MapData
     */
    totalGames: number;
    /**
     * 
     * @type {Array<MapPlayerWL>}
     * @memberof MapData
     */
    playerStats: Array<MapPlayerWL>;
    /**
     * 
     * @type {Array<MapGeneralWL>}
     * @memberof MapData
     */
    generalStats: Array<MapGeneralWL>;
}

/**
 * Check if a given object implements the MapData interface.
 */
export function instanceOfMapData(value: object): value is MapData {
    if (!('mapName' in value) || value['mapName'] === undefined) return false;
    if (!('totalGames' in value) || value['totalGames'] === undefined) return false;
    if (!('playerStats' in value) || value['playerStats'] === undefined) return false;
    if (!('generalStats' in value) || value['generalStats'] === undefined) return false;
    return true;
}

export function MapDataFromJSON(json: any): MapData {
    return MapDataFromJSONTyped(json, false);
}

export function MapDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): MapData {
    if (json == null) {
        return json;
    }
    return {
        
        'mapName': json['mapName'],
        'totalGames': json['totalGames'],
        'playerStats': ((json['playerStats'] as Array<any>).map(MapPlayerWLFromJSON)),
        'generalStats': ((json['generalStats'] as Array<any>).map(MapGeneralWLFromJSON)),
    };
}

export function MapDataToJSON(json: any): MapData {
    return MapDataToJSONTyped(json, false);
}

export function MapDataToJSONTyped(value?: MapData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'mapName': value['mapName'],
        'totalGames': value['totalGames'],
        'playerStats': ((value['playerStats'] as Array<any>).map(MapPlayerWLToJSON)),
        'generalStats': ((value['generalStats'] as Array<any>).map(MapGeneralWLToJSON)),
    };
}