// src/api/models/MapStatsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MapData } from './MapData';
import {
    MapDataFromJSON,
    MapDataFromJSONTyped,
    MapDataToJSON,
    MapDataToJSONTyped,
} from './MapData';

/**
 * 
 * @export
 * @interface MapStatsResponse
 */
export interface MapStatsResponse {
    /**
     * 
     * @type {Array<MapData>}
     * @memberof MapStatsResponse
     */
    maps: Array<MapData>;
}

/**
 * Check if a given object implements the MapStatsResponse interface.
 */
export function instanceOfMapStatsResponse(value: object): value is MapStatsResponse {
    if (!('maps' in value) || value['maps'] === undefined) return false;
    return true;
}

export function MapStatsResponseFromJSON(json: any): MapStatsResponse {
    return MapStatsResponseFromJSONTyped(json, false);
}

export function MapStatsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MapStatsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'maps': ((json['maps'] as Array<any>).map(MapDataFromJSON)),
    };
}

export function MapStatsResponseToJSON(json: any): MapStatsResponse {
    return MapStatsResponseToJSONTyped(json, false);
}

export function MapStatsResponseToJSONTyped(value?: MapStatsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'maps': ((value['maps'] as Array<any>).map(MapDataToJSON)),
    };
}